Abstract UserActionEmailCongressperson components
