+++ src/components/app/userActionFormEmailCongressperson/common/constants.ts
@@ -1,2 +1,7 @@
 export const ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON =
   'User Action Form Email Congressperson'
+
+export const SectionNames = {
+  EMAIL: 'email',
+  SUCCESS: 'success',
+}
