+++ src/components/app/userActionFormEmailCongressperson/dialog.tsx
@@ -1,62 +1,65 @@
 'use client'
 
-import { Suspense, useEffect, useState } from 'react'
+import { Suspense } from 'react'
+import dynamic from 'next/dynamic'
 
 import { UserActionFormDialog } from '@/components/app/userActionFormCommon/dialog'
-import { ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON } from '@/components/app/userActionFormEmailCongressperson/constants'
-import { LazyUserActionFormEmailCongressperson } from '@/components/app/userActionFormEmailCongressperson/lazyLoad'
+import { ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON } from '@/components/app/userActionFormEmailCongressperson/common/constants'
 import { UserActionFormEmailCongresspersonSkeleton } from '@/components/app/userActionFormEmailCongressperson/skeleton'
-import { UserActionFormEmailCongresspersonSuccess } from '@/components/app/userActionFormEmailCongressperson/success'
 import { FormFields } from '@/components/app/userActionFormEmailCongressperson/types'
-import { UserActionFormSuccessScreen } from '@/components/app/userActionFormSuccessScreen'
-import { dialogContentPaddingStyles } from '@/components/ui/dialog/styles'
 import { useApiResponseForUserFullProfileInfo } from '@/hooks/useApiResponseForUserFullProfileInfo'
-import { useCountryCode } from '@/hooks/useCountryCode'
 import { useDialog } from '@/hooks/useDialog'
-import { cn } from '@/utils/web/cn'
+import { SupportedCountryCodes } from '@/utils/shared/supportedCountries'
+
+const UserActionFormEmailCongressperson = dynamic(
+  () =>
+    import('@/components/app/userActionFormEmailCongressperson').then(
+      mod => mod.UserActionFormEmailCongressperson,
+    ),
+  {
+    loading: () => (
+      <UserActionFormEmailCongresspersonSkeleton countryCode={SupportedCountryCodes.US} />
+    ),
+  },
+)
+
+interface UserActionFormEmailCongresspersonDialogProps {
+  children: React.ReactNode
+  defaultOpen?: boolean
+  initialValues?: FormFields
+  countryCode: SupportedCountryCodes
+}
 
 export function UserActionFormEmailCongresspersonDialog({
   children,
   defaultOpen = false,
   initialValues,
-}: {
-  children: React.ReactNode
-  defaultOpen?: boolean
-  initialValues?: FormFields
-}) {
+  countryCode,
+}: UserActionFormEmailCongresspersonDialogProps) {
   const dialogProps = useDialog({
     initialOpen: defaultOpen,
     analytics: ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON,
   })
-  const countryCode = useCountryCode()
   const fetchUser = useApiResponseForUserFullProfileInfo()
-  const [state, setState] = useState<'form' | 'success'>('form')
   const { user } = fetchUser.data || { user: null }
-  useEffect(() => {
-    if (!dialogProps.open && state !== 'form') {
-      setState('form')
-    }
-  }, [dialogProps.open, state])
+
+  if (fetchUser.isLoading) {
+    return (
+      <UserActionFormDialog {...dialogProps} padding={false} trigger={children}>
+        <UserActionFormEmailCongresspersonSkeleton countryCode={countryCode} />
+      </UserActionFormDialog>
+    )
+  }
 
   return (
     <UserActionFormDialog {...dialogProps} padding={false} trigger={children}>
       <Suspense fallback={<UserActionFormEmailCongresspersonSkeleton countryCode={countryCode} />}>
-        {fetchUser.isLoading ? (
-          <UserActionFormEmailCongresspersonSkeleton countryCode={countryCode} />
-        ) : state === 'form' ? (
-          <LazyUserActionFormEmailCongressperson
-            initialValues={initialValues}
-            onCancel={() => dialogProps.onOpenChange(false)}
-            onSuccess={() => setState('success')}
-            user={user}
-          />
-        ) : (
-          <div className={cn(dialogContentPaddingStyles, 'h-full')}>
-            <UserActionFormSuccessScreen onClose={() => dialogProps.onOpenChange(false)}>
-              <UserActionFormEmailCongresspersonSuccess />
-            </UserActionFormSuccessScreen>
-          </div>
-        )}
+        <UserActionFormEmailCongressperson
+          countryCode={countryCode}
+          initialValues={initialValues}
+          onCancel={() => dialogProps.onOpenChange(false)}
+          user={user}
+        />
       </Suspense>
     </UserActionFormDialog>
   )
