+++ src/components/app/userActionFormEmailCongressperson/homepageDialogDeeplinkWrapper.tsx
@@ -1,17 +1,14 @@
 'use client'
 
-import React, { Suspense, useEffect, useState } from 'react'
+import React, { Suspense, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 
 import { GeoGate } from '@/components/app/geoGate'
 import { UserActionFormActionUnavailable } from '@/components/app/userActionFormCommon/actionUnavailable'
-import { UserActionFormEmailCongressperson } from '@/components/app/userActionFormEmailCongressperson'
-import { ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON } from '@/components/app/userActionFormEmailCongressperson/constants'
+import { ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON } from '@/components/app/userActionFormEmailCongressperson/common/constants'
 import { UserActionFormEmailCongresspersonSkeleton } from '@/components/app/userActionFormEmailCongressperson/skeleton'
-import { UserActionFormEmailCongresspersonSuccess } from '@/components/app/userActionFormEmailCongressperson/success'
 import { FormFields } from '@/components/app/userActionFormEmailCongressperson/types'
-import { UserActionFormSuccessScreen } from '@/components/app/userActionFormSuccessScreen'
-import { dialogContentPaddingStyles } from '@/components/ui/dialog/styles'
+import { USUserActionFormEmailCongressperson } from '@/components/app/userActionFormEmailCongressperson/us'
 import { trackDialogOpen } from '@/components/ui/dialog/trackDialogOpen'
 import { useApiResponseForUserFullProfileInfo } from '@/hooks/useApiResponseForUserFullProfileInfo'
 import { useCountryCode } from '@/hooks/useCountryCode'
@@ -19,7 +16,6 @@ import { useEncodedInitialValuesQueryParam } from '@/hooks/useEncodedInitialValu
 import { usePreventOverscroll } from '@/hooks/usePreventOverscroll'
 import { DEFAULT_SUPPORTED_COUNTRY_CODE } from '@/utils/shared/supportedCountries'
 import { getIntlUrls } from '@/utils/shared/urls'
-import { cn } from '@/utils/web/cn'
 
 function UserActionFormEmailCongresspersonDeeplinkWrapperContent() {
   usePreventOverscroll()
@@ -28,8 +24,8 @@ function UserActionFormEmailCongresspersonDeeplinkWrapperContent() {
   const router = useRouter()
   const countryCode = useCountryCode()
   const urls = getIntlUrls(countryCode)
-  const [state, setState] = useState<'form' | 'success'>('form')
   const { user } = fetchUser.data || { user: null }
+
   const [initialValues, loadingParams] = useEncodedInitialValuesQueryParam<FormFields>({
     address: {
       description: '',
@@ -43,21 +39,17 @@ function UserActionFormEmailCongresspersonDeeplinkWrapperContent() {
     trackDialogOpen({ open: true, analytics: ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON })
   }, [])
 
-  return fetchUser.isLoading || loadingParams ? (
-    <UserActionFormEmailCongresspersonSkeleton countryCode={countryCode} />
-  ) : state === 'form' ? (
-    <UserActionFormEmailCongressperson
+  if (fetchUser.isLoading || loadingParams) {
+    return <UserActionFormEmailCongresspersonSkeleton countryCode={countryCode} />
+  }
+
+  return (
+    <USUserActionFormEmailCongressperson
+      countryCode={countryCode}
       initialValues={initialValues}
       onCancel={() => router.replace(urls.home())}
-      onSuccess={() => setState('success')}
       user={user}
     />
-  ) : (
-    <div className={cn(dialogContentPaddingStyles, 'h-full')}>
-      <UserActionFormSuccessScreen onClose={() => router.replace(urls.home())}>
-        <UserActionFormEmailCongresspersonSuccess />
-      </UserActionFormSuccessScreen>
-    </div>
   )
 }
