+++ src/components/app/userActionFormEmailCongressperson/us/campaignMetadata.ts
@@ -3,7 +3,7 @@ import { EmailActionFormValues } from '@/components/app/userActionFormEmailCongr
 import {
   getFullNameSignOff,
   GetTextProps,
-} from '@/components/app/userActionFormEmailCongressperson/emailBodyUtils'
+} from '@/components/app/userActionFormEmailCongressperson/common/emailBodyUtils'
 import { USUserActionEmailCampaignName } from '@/utils/shared/userActionCampaigns/us/usUserActionCampaigns'
 import { YourPoliticianCategory } from '@/utils/shared/yourPoliticianCategory'
